Initial Commit
This is the start of the Grocery Store Price Tracker project. This is
supposed to be a project that tracks and stores grocery store prices
over a long period of time. It will be a full stack application that
stores archived prices for groceries at major Canadian grocery chains,
using Spring Boot on the back-end, and using selenium with Java to
scrape the websites involved. The front-end will likely be done with
REACTjs. The database is yet to be decided, but it may be MySQL, and the
database may be interfaced with an Object Relational Mapping. Currently
I am working on the scraping of the sites for data, studying the HTML of
different grocery chains. The current object oriented design is to have
an interace in GroceryStorePriceScraper.java with methods that all
iterators will implement. The iterators will scrape through the prices
over time and return a hashmap with price and product data. This will be
put into the database with a timestamp for when the data was gathered.