Added to the loblaws iterator, and made pom.xml changes
I added to the loblaws iterator, as the original code was just placeholder
code.
1) I added to the LoblawsIterator.java file the code for opening the
webdriver at a given url, going to the store locator and typing "Canada"
to get all locations in Canada, and then getting each city. For each city,
I stored the city as a key in a hashmap, with its value being
Boolean(true). This serves to get all city locations for all stores, so
that a store can be selected for each city and the prices be scraped for
all cities. The reason for this is because prices at stores in
different cities can be different. This ensures that all price
discrepancies are accounted for across different cities. The plan will
be to type in each city in the store locator, and select the first store
that shows in that city. For that city, prices will be scraped from that
store only, as prices are not expected to change across different stores
in the same city.
2) I added changes to the pom.xml file to include selenium and other
   libraries related to selenium.
3) Since the LoblawsIterator.java code will be scraping the website at a
   given url, there are a lot values that would best be taken from a
   .properties file (such as the url address, css selector values of
   various html elements, etc.). Otherwise, they would be hard-coded
   into the code itself, which requires more maintenance (i.e. the code
   has to repeatedly be compiled). The hard-coded values are stored in
   config/loblaws.properties.
4) I added the target/ directory in the .gitignore file, as that is the
   output of running maven commands to compile and package the project.
5) I used the dos2unix command on all of the files in this commit due to
   line endings being different. So some files have no change but are in
   this commit for that reason. Other files have the entire old lines
   plus the new lines being in the commit for the same reason.