Starting to add code to get prices for each township
I have started to add code to get prices for each township. So the
workflow will be that you redirect to the main nofrills.ca website, and
change the location of the store pickup to be one store for each city in
the cities gathered. For each string representing a unique city that was
added in the string boolean hashmap, we enter it in the search bar for
the store locator. Then we select the location to be the first store
that comes in the results. Then we click the top menu above (Where
Grocery is found), and then click each of the sub-items under this menu
one-by-one. From there, in the subgenre of groceries, we click on each
subtype for the subgenre of groceries, and it opens up what I believe to
be a Bootstrap CSS accordion. One of the buttons has the text labelled
"See All", which is then clicked. The next step from here is to scrape
all of the prices that are listed here. There is a pagination list of
prices (consisting of one or more pages).

Some items will be skipped. For example, under the grocery menu heading,
the items in the sub-menu titled "Flower Shop" and "Seasonal Shop" will
not be parsed. This because "Flower Shop" is not relevant for our
purposes, and "Seasonal Shop" is redundant. All the items listed under
this category are in other categories. It just covers the items that are
in-season. There are helper functions to determine if these items are to
be skipped based on the CSS selector value, and the listed items to be
skipped. For all of the added code, additions had to be made to the
loblaws.properties file, hence all of the additions to the
loblaws.properties file.