Added code to scrape over individual prices of products in a store
I have started to add code to scrape over product and price data for
individual products for an area of groceries for a store in question.
I added a helper function (fluentWaitTillVisibleandClickable) to wait
till the element is present and visible, and then return the element.
I have a boolean function to return true if this the element is present
and visible (elementExistsAndIsInteractable). There are also other
helper functions to change the numbers in StringBuilder objects and
return the new StringBuilder object (i.e. changing the "1" to "2"
in "nth-child(1)"). Then the main function added in this commit
(scrapeAllPrices) leverages these functions to scrape product
information for all products showing for the current subcategory of
product. For example, if the current subcategory of product is "Fresh Fruits"
or "Fresh Vegetables" (or any other category), the function scrapes
product and price data for all products listed under this category. From
here, the scraper will iterate over all categories and subcategories of
products and gather product information for all products listed. It will
do this over all cities and townships where a NoFrills store exists.
Currently, the product information is stored in a HashMap, and these
HashMaps are appended to an instance of ArrayList<HashMap<String, String>>.
This ArrayList is currently not returned, but it will likely be returned
in the form of one HashMap over each iteration of the Iterator (as the
design of this will be modified to be an iterator). This added code
required also adding more information in the loblaws.properties file
(more variables with values of relevant CSS selectors). This is why
there were changes to the loblaws.properties file.