Added more method overloads for helper functions
Added more method overloads for helper functions that changed the index
inside a selector (changing the digit in "nth-child(1)" to be one more,
or manually changing the digit to be a number passed in). The new method
overloads for these helper function have a starting index that you can
start searching the round brackets from. For example, you can  start
searching from the third character by passing in 2 as the starting
index, or the tenth character by passing in  9 as the starting index.