Added the code to scrape over all product subcategories
I added the code to iterate within all for a subcategory of the
subcategories under the main menu subcategories. To illustrate, you have
main menu categories (A) -> subcategories under main menu category (B) ->
subcategories within subcategory under main menu category (C). The code now
will iterate over all subcatgories in (C) (described in previous
sentence). So for a store, under the Fruits and Vegetables category
(which falls under the category of Groceries), there are subcategories
of products under the Fruits and Vegetables cateogry. The code iterates
over all those subcategories of products under the Fruits and Vegetables
category. The next step from here is to iterate over all subcategories
listed under each main menu category (described above as "(B)").

There were also instances where under a page with products listed, there
was no nav element with a pagination list in it. This is because there
was only one page of listed products. So I made sure to account for this
as well.

I changed the key values for the key, value pairs that were being put
into each hashmap of product information.

I also added changes to the loblaws.properties file to account for new
selectors that were required. I also had to change a number value that
would be needed to change the number index in a certain set of selectors
at a certain spot.