Added code to parse product info for each product in separate tab
I added code to parse product information for each product in a separate
tab. For each product, the link to its product page is opened in a
separate tab and the information is scraped from there. There are some
elements that are always there, and there are some elements that do not
show up always (i.e. brand name and package size are not always there).
So the code has to account for this, as well as the unit price data not
always being there. Currently, if the product is on sale at that
location, the sale price is taken as the current price. This can be
changed to have the sale price and regular price be separate keys and
values in the HashMap, depending on what elements and selectors are
present. For product information to be opened in a new page, new
selectors need to be added to the loblaws.properties file, hence the
changes to the config/loblaws.properties file.