Added ability to iterate over subcategories in grocery category
I added the ability to iterate over all subcategories in the grocery
category, except for certain ones that are to be ignored (specified in
the loblaws.properties file). Earlier, the code just iterated over one
subcategory under the grocery category. The next step is to iterate over
the main categories. Visually you can view it as fourth-degree nested
iteration. First we iterate over the main categories. In each iteration
(which we'll call iteration A), we iterate over the subcategories of
the main categories (which we'll call iteration B). Then we iterate
over the subcategories under the subcategories of the main category
(which we'll call iteration C), and then over the products listed
under those subcategories of subcategories of main categories (which
we'll call iteration D). So far the code has implemented nested
iterations of iterations B, C, and D. What's left is implementing nested
iterations of B, C and D inside each iteration of A (iteration A still
has to be implemented). This requires that we also change some selectors
in the loblaws.properties file, and add an integer used in the code for
incrementing digits in a CSS selector (represented by a StringBuilder).
Hence, changes were added to the loblaws.properties file.

I added a helper function to get variables out of the .properties file
that are integers, and try to return the parsed integer from the value
stored in the .properties file (which is a string).

I also added code to scroll certain submenu items into view, as
sometimes the submenu items are not visible, and moving to and clicking
the element causes errors. So I added some lines to scroll the element
into view before clicking the element.