Added functionality to scrape over all main categories
In the previous commit message, you'll note that iteration A had to be
implemented. Now it has been implemented. So the nested quadruple for
loop mentioned in the previous commit message is implemented (iteration
D nested in C, which is nested in B, which is nested in iteration A).
I also changed the helper method which checks the main menu items that
should be ignored. Before the method would look at the current element
and return false if the element's text was in a list of elements from
a variable in the config file. Now the method returns true if the
element's text is in a comma-separated list of phrases from the same
variable in the config file. In other words, the list of elements now
represents what to ignore instead of what not to ignore. The code in the
method has been changed for this, as well as the value of the variable
in the loblaws.properties file.