Changed selector value that was not being found
When you get to the point where the indivudal products are listed, you
see breadcrumbs at the top of the page representing the path to that
page. The end of the breadcrumb represents the current page that you are
on. I was originally trying to get the immediate parent of the current
page using a CSS selector containing "li:nth-child(2)". This would not
always work because the list of ancestor pages for different pages of
products would not always be the same list (i.e. they vary in lengths).
So instead I changed "li:nth-child(2)" to "li:nth-last-child(2)". This
ensures that we always get the link in the breadcrumb for the immediate
parent page of the current page.