Refactored code to add helper method
Refactored code to add a helper method that will scroll an element into
view, move the mouse over the element, pause and then click on the
element. I also added a helper method that will scroll the element into
view, move the cursor to the element, pause for a given number of
milliseconds, click the element, and then pause for another given number
of milliseconds. I used this to replace code in places where I used
an ActionChain instead.