Refactored code, fixed compilation error, and added maven dependency
Refactored java code to have the hasNext and next methods placed at the
bottom. I fixed a compilation error related to the use of the type
"webElement" (which isn't declared) and replaced it with "WebElement"
(the class type for selenium web elements).

I added a new maven dependency for the xstream library (to serialize
objects to xml and vice versa). I added this because for the loblaws
iterator, the plan was to scrape the data and then (with the next()
method) return the hashmap for product information on one product. The
problem is that it requires for me to refactor the entire codebase
around the next() and hasNext() methods (to implement the iterator
interface). It is a lot of work to do this, so instead the plan is to
scrape all the information and put it in an xml file, and then parse the
xml file. From there, the data can be stored in another format (i.e. in
a database). There will be a method to put all the scraped data in an
xml file, and the next() method will return hashmaps from the xml file.
This is only for the loblaws iterator. Other iterators will center
around the next() and hasNext() methods.