Modified the code to use the Java STAX API instead of XStream
1) Modified the code to use STAX API instead of XStream

Summary - I have modified the code to use the Java STAX API via
XMLEventWriter instead of XStream. This is because the STAX API
is more memory efficient.

Explanation: The STAX API loads the xml file line by line into memory
and writes out to the xml file incrementally (i.e. writes out to the xml
file tag by tag). This is useful when your xml file is projected to be
thousands of lines long. In this case, you are scraping price data for
each product in each category, in each city in Canada where there is a
NoFrills store. Within a single store, this amounts to thousands of
products. Doing this for one store location for each city in Canada
means that the resulting XML file with all of the data will be hundreds
of thousands, if not millions of lines long. The entries in the XML tags
will be database entries, meaning that the XML file will be read into
memory first. This is not feasible when the XML files are large. To
account for this new change, changes were made to
src/main/java/iterators/App.java as well.

2) Solved an issue where a submenu item was not being clicked.

A submenu item was not being clicked, since it wasn't being scrolled
into view. So a fix was made to force the element to be scrolled into
view (using scrollIntoView(true) only if scrollIntoView(false) didn't
work first).

3) I added changes in the loblaws.properties file to account for CSS
   selector changes. I also added changes to the properties file to
   account for using the STAX API.

4) I added a change to the .gitignore file to ignore xml files in the
   config directory.