Added code to append to xml file
I changed the directory in which the xml file will be, and added code to
append to the xml file if it already exists with a product_records tag.
If the xml file exists with a product_records tag and product_record
entries inside the product_records tag, then a new product_records tag
will be created. product_record tag entries will be inside this new
tag.