Name change for a method, and a documentation update
I made a name change for the getURL method to be the loadXML method.
I changed it to have return type void as its job is just to load
entries into the XML file. I updated the use of this method in
src/main/java/iterators/App.java. I updated the documentation and
method name for openProductXmlStream to be openProductXmlOutputStream.
It will take care of only the output stream to the xml file. Another
method will take care of opening the input stream.