Moved the xml parsing code to a separate class
I moved the xml parsing code to a separate class to aim for
high-cohesion and low-coupling splits. This allows for the code to
be re-usable by other iterators for other stores (i.e. in future an
iterator could be made for Walmart, another for FreshCo, etc.).
The next step is to finish adding documentation and add a couple of
other features to the Loblaws Iterator.