Added documentation for the LoblawsIterator and XMLParser classes
I added documentation for all methods in the LoblawsIterator and
XMLParser classes. I added some code for the methods that change
digits within round brackets in CSS selectors. The code would
catch the case where the opening bracket isn't found, and return a
StringBuilder object with the string being the original CSS selector.