Added code for Document Object Model (DOM) XML parsing
Context:
In the case that a time limit is specified, parsing every product for
every city will not be feasible. Parsing every product for one city
alone will take probably a few hours. In this case, an xml file will be
made with tags with the text content in each tag being a city to parse
grocery prices in. Each time prices are parsed in a city, that tag is
removed from the parsed DOM in the XML. Once the time limit is up, this
xml file is saved with the tags of remaining cities that you have to
parse prices in (the modified DOM is written to the XML file). The
next time you run the iterator, it will read in this xml file and
pick up where you had left off (the first tag). So if there are 300
cities to parse during a certain time limit, the iterator will go
through the as many cities as possible and delete those tags from the
DOM in the process. If there are 293 cities left once time is up, then
the new DOM with 293 city entries is written back to the XML file. The
next time you start up the iterator, it will progress through the
remaining 293 entries (as much as possible if there is again a specified
time limit).

Example XML Change:
cities.xml is created and the following is written to it during the
first time interval.
<cities>
	<city>Toronto</city>
	<city>Ottawa</city>
	<city>Montreal</city>
</cities>

The iterator only scrapes prices in toronto during the time limit. The
following is written back into cities.xml.
<cities>
	<city>Ottawa</city>
	<city>Montreal</city>
</cities>

The next time the iterator is started up, this file is read in again
and the iterator progresses through the remaining cities.

I also added supporting code to implement this in LoblawsIterator.java,
and removed redundant/unnecessary code.

To Do:
* Finish implementing the code in LoblawsIterator.java to implement this
XML solution for cities left to parse.
* Account for the case where in a city not all products have been
  parsed.