Added logic for time limits and for saving and resuming progress
I added logic to set a time limit, and to have the program run until the
time limit is reached. It will then save the new results to the product
information xml file and save the information of cities left to scrape
in the xml file with city information. I added logic to check if the xml
file with cities left to scrape still exists, so that it can continue
from the cities that are listed in this file. I also refactored some
code out of the getAllCities function, since it has to be executed
regardles if getAllCities is executed. I modified getAllCities to
account for the case where a time limit is specified and you need to
create and XML file with data on cities that have to be scraped. For
this to happen, I had to make the createXMLNode methods in the XMLParser
class public. I also had to remove the XMLEventWriter argument (the
former first argument) to both method definitions. I also set the
removeChild method on this.doc to be on this.root_node instead in the
DOMParser class, since this was causing an error with the node not being
found.