Added code to delete the cities-related XML file once all cities parsed
Added code such that once all cities are parsed in the cities-related
XML file, the XML file is deleted. This ensures that there are no cities
left, the program starts fresh the next time it is ran and scrapes all
cities. This also ensures that once all cities are parsed and the next
time you run the program with a time limit, a new XML file is created.
In this new XML file, all the cities are added again from the beginning
(and they are parsed again one by one).