Added more robustness to methods gathering product information
I added more robustness to methods gathering product information, such
that if the page fails to load the program does not terminate due to an
uncaught TimeoutException or a NoSuchElementException. If the page fails
to load, the new tab is closed and an empty hashmap is returned. The
method to scrape product information is wrapped within another method.
If scraping the product information fails, the wrapper method has a
certain number of retries that it uses till it returns an empty hashmap
itself. If an empty hashmap is returned, it is not added to the XML
file. I also mistakenly added remaining_cities.xml, and will remove it
in the next commit.