Fixed bug relating to adding of categories and subcategories
In the addSubCategories method and the addCategories method, I would add
the new categories to the beginning of the array instead of the end.
Then the categories would be removed one by one from index 0. However,
that is not the order in which the categories would be scraped, and by
which you would go through the categories and subcategories. So you
would be removing the wrong categories/subcategories, and the resulting
XML files would show the wrong categories/subcategories to continue from
in the next run. This bug was due to adding new categories/subcategories
at the beginning of the ArrayLists, instead of at the end. That way the
order of the categories and subcategories in the ArrayLists would match
the order in which you scrape through them. So when you remove them and
put the results in the XML files, the categories and subcategories to
continue from are the right ones.