Added code to continue scraping on same city if city is not fully done
Note - in the paragraphs below, the current node and previous node refer
to this.current_node and this.previous_node respectively (in the code
for src/main/java/iterators/xml/DOMParser.java).

The previous commit message talks about how in each run the city
changes, regardless of whether or not scraping was finished for one
city. This was because of the DOMParser's next() method returning the
text in the next node, but deleting that same node. So the scraping for
one city could only be done on one run. Now the code handles different
cases differently. When the first run of the DOMParser's next() method
starts, the first node is the current node. On the first call to
next(), the text in the first node is returned and the first node
becomes the previous node, while the second node becomes the current
node. When the second next() call starts, the second node is the
current node. When the second next() call completes the second node
becomes the previous node, and the third node becomes the current node.
The return value of the second next() call is the text in the second
node. In the second call of next(), the first node also gets deleted. In
the third call of next(), the second node is the previous node and gets
deleted. The text in the third node is returned and at the end of the
third next() call, the third node becomes the previous node, while the
fourth node becomes the current node.

This continues until we get to the last call to the next method. Here we
have the last two nodes left, and the previous node is the second last
node while the current node is the last node. Here, the second last node
gets deleted, the text for the last node is returned, and
this.current_node gets set to equal this.previous_node. There is a
conditional that if this.current_node equals this.previous_node, then
this.current_node should also be deleted. So in the last call to next(),
the text for the last node is returned, and the last two nodes are
deleted (i.e. all remaining nodes are deleted). This accounts for the
case where you are scraping product information in one city, and you do
not finish scraping all products in that city within the allotted time.
Now, you can continue scraping product information on the same city.