Refactored the scrapeAllPrices method to be made up of helper methods
The scrapeAllPrices method (formerly named the scrapePrices method) was
becoming too long. So I split the method into smaller helper methods.
The parts of the method where you close the sign in button and select
the correct filters are offset into different helper methods. The newly
defined scrapePrices method (whos function body is empty in this commit)
will be responsible for scraping all product entries on the page, and
then putting it into the xml file. Iterating over all pages will be done
in the getAllPrices method.