Added code to parse all products for first page of search results
I added code to the MetroIterator class to parse through all products
(product containers) on the first page of the search results. It
required adding new static utility methods in the WebElementOperations
class. The first one is to get the innerText property for an element
using JavaScript. The second one is to check if an element exists using
JavaScript. But instead of starting from the root document (using
document.querySelector), it starts from a specific element (using
element.querySelector). These methods help with determing if a specific
tag with a certain unique class value exists within the container
element for a product. The innerText property helps get the text from
the element and its descendants (the text as you would read it on the
screen). Both of these are used to parse through each product container
and scrape all possible information from the container to put in the XML
file. So far the code only scrapes all products on the first page. The
next step is to iterate through all pages of results (using the
pagination, and our previously defined methods) to put them in the XML
file.