Added code to MetroIterator to change store locations for all cities
I added code to the MetroIterator class to change store locations and
then scrape the data afterwards. This would be done after first
gathering all cities into a XML file containing cities left. However,
Cloudflare seems to have caught on to the workflow I'm using. So
although the code for the MetroIterator is done, none of it seems to be
of any use. I will next work on making an iterator for Walmart, or start
with creating the Database as well as the REST API back-end using the
Spring Framework.