Added code for a MySQL database client
I added code for a MySQL database client that will act as the client to
the database. It will take the information in the XML files and insert
it into the database tables. It will provide some abstraction over the
details of writing entries to the database and reading from the
database. The plan is that once the information is scraped, the items
from the xml files are taken and then reorganized into xml files for the
different database tables (i.e. an XML file for entries in the "fruits
and vegetables" table, another for the "dairy and eggs" table, etc.).
From there, the items will be taken from the XML file and written into
the database one by one. The DatabaseClient class may provide an
iterator abstraction over querying the database as well, if there are
many results.