Added code to separate XML results into different files for food groups
I added code to separate XML results into different files for the
different categories of food groups. The categories of food groups are
based on the table names in the database. There are different XML
parsers for the different XML files (one for each file representing a
food group). Once the scraping is done, the results are iterated over
and based on the category, they are put into the different XML files.

I also added a query method to the DatabaseClient that allows you to to
specify a query for the DatabaseClient to run. You pass in the SQL code
you wish to run, and the DatabaseClient runs the query for you. So far
it has only insert statements in mind, which is why the method has a
return type of void. In the future more will be done to the
DatabaseClient class. For example, accommodations will be made to
include SELECT statements. Another important point to look into is
handling the case where the results returned from the SELECT statement
span more memory than is available in your machine's RAM section.