Added code to parse wikipedia for most populated cities in Canada
I added code to search a wikipedia link for the most populated cities in
Canada. This new code would allow for us to scrape Canada's most
populated cities and put it in an XML file for the LoblawsIterator class
to use. Assuming that the XML files representing cities are the same
with the same root tags and mapping tags, we have that the xml file
created by the PopularCities class can then be used by the
LoblawsIterator's loadXML method. This purpose of this is to reduce the
number of cities for which prices need to be scraped.

This will help reduce the number of entries in the tables of our
database. This is important as for NoFrills stores, the total number
of products scraped in each city is between 5000 and 15000. We can
specify the limit of the number of cities to parse as a parameter to the
PopularCities constructor. The value itself is stored in a .properties
configuration file. Currently we parse the 10 most populated cities.