Added code to scrape information about most popular cities
Before the PopularCities class had some code to parse the most populated
cities, but the issue was that the driver would not start up. So I added
some code to specify the driver path, and that helped start up the
driver for the class. From there, the issue was that all the cities
would be scraped, when it should have been no more than the limit passed
in. This was due to a logic error in the while loop so I fixed the while
loop to account for this. Now the code worked, and the 10 most populated
cities were successfully read in from the link I specified in the
.properties file.

I forgot about the case where the file already exists, in which case no
scraping should be done. So I added code to check if the XML file
associated with the XMLParser currently exists. This required adding an
xmlFileExists method to check if the XML file exists. But when I tried
running the method it failed because a property in the class had not yet
been instantiated. This was because the method to open the output stream
had not yet been called (which would set the property that was needed
for the xmlFileExists method). So I moved the necessary code (that would
set the property in question) into another method that would be called
immediately in the constructor for XMLParser. This would ensure that we
can check that the XML file in question exists or not, and if it exists
then we do not scrape information on the most popular cities. If it
doesn't exist, we scrape information on the most popular cities in
Canada.

There was one other case I didn't account for. When changing the
location where you scrape prices, you may enter a location that returns
no matches in that city. So I added code to ensure that when you search
stores for a city, the first result you get is in the city in question.
Otherwise, what currently happens is that if no locations in that city
are found, then the stores in the nearest townships appear first. The
first location would be selected, and you would be scraping prices in a
city that was not the one you wanted to be scraping prices for.